complex keybind and removed LEFT keybind
